requirements.txt
openai-whisper
python-telegram-bot
pydub
ffmpeg-python
python-dotenv